API

Ark's public API.

World

Map

Entities

Components

Queries

Resources

Batch

Index

World

The World is the central data storage for Entities, Components and Resources.

Ark.World — Type

World{CS<:Tuple,CT<:Tuple,N}

The World is the central ECS storage.

Ark.is_locked — Function

is_locked(world::World)::Bool

Returns whether the world is currently locked for modifications.

Map

Maps allow access and manipulation capabilities for entities. Note that the same capabilities are also accessible through the World, but the maps API may be more ergonomic in some cases. See also the Components-related API.

Ark.Map — Type

Map{CS, N}

A component mapper for N components.

Ark.@Map — Macro

@Map(world::World, comp_types::Tuple)

Macro version of Map for ergonomic construction of component mappers.

Arguments

world: The World instance to map components from.

comp_types::Tuple: A tuple of component types, e.g. (Position, Velocity).

Example

@Map(world, (Position, Velocity))

Base.getindex — Method

Base.getindex(map::Map, entity::Entity)

Get the Map's components for an Entity.

Base.setindex! — Method

Base.setindex!(map::Map, values::Tuple, entity::Entity)

Sets the values of the Map's components for an Entity. The entity must already have all these components.

Entities

Entities are the "game objects" or "model entities". An entity if just an ID with a generation, but Components can be attached to an entity.

Ark.Entity — Type

Entity

Entity identifier.

Ark.zero_entity — Constant

const zero_entity::Entity

The reserved zero Entity value.

Ark.new_entity! — Function

new_entity!(world::World, values::Tuple)::Entity

Creates a new Entity with the given component values. Types are inferred from the values.

new_entity!(map::Map, comps::Tuple)::Entity

Creates a new Entity with length(comps) components.

Ark.new_entities! — Function

new_entities!(world::World, n::Int, defaults::Tuple; iterate::Bool=false)::Union{Batch,Nothing}

Creates the given number of Entity, initialized with default values. Component types are inferred from the provided default values.

If iterate is true, a Batch iterator over the newly created entities is returned that can be used for initialization.

Arguments

world::World: The World instance to use.

n::Int: The number of entities to create.

defaults::Tuple: A tuple of default values for initialization, like (Position(0, 0), Velocity(1, 1)).

iterate::Bool: Whether to return a batch for individual entity initialization.

new_entities!(world::World, n::Int, comp_types::Tuple{Vararg{Val}})::Batch

Creates the given number of Entity.

Returns a Batch iterator over the newly created entities that should be used to initialize components. Note that components are not initialized/undef unless set in the iterator!

For a more convenient tuple syntax, the macro @new_entities! is provided.

Arguments

world::World: The World instance to use.

n::Int: The number of entities to create.

comp_types::Tuple: Component types for the new entities, like Val.((Position, Velocity)).

Ark.@new_entities! — Macro

@new_entities!(world::World, n::Int, comp_types::Tuple{Vararg{Val}})::Batch

Creates the given number of Entity.

Returns a Batch iterator over the newly created entities that should be used to initialize components. Note that components are not initialized/undef unless set in the iterator.

Macro version of new_entities! for ergonomic construction of component mappers.

Arguments

world::World: The World instance to use.

n::Int: The number of entities to create.

comp_types::Tuple: Component types for the new entities, like (Position, Velocity).

Ark.remove_entity! — Function

remove_entity!(world::World, entity::Entity)

Removes an Entity from the World.

Ark.is_alive — Function

is_alive(world::World, entity::Entity)::Bool

Returns whether an Entity is alive.

Ark.is_zero — Function

is_zero(entity::Entity)::Bool

Returns whether an Entity is the zero entity.

Components

Components contain the data associated with Entities

Ark.get_components — Function

get_components(world::World, entity::Entity, comp_types::Tuple)

Get the given components for an Entity.

For a more convenient tuple syntax, the macro @get_components is provided.

Example

pos, vel = get_components(world, entity, Val.((Position, Velocity)))

Ark.@get_components — Macro

@get_components(world::World, entity::Entity, comp_types::Tuple)

Get the given components for an Entity.

Macro version of get_components for more ergonomic component type tuples.

Example

pos, vel = @get_components(world, entity, (Position, Velocity))

Ark.has_components — Function

has_components(world::World, entity::Entity, comp_types::Tuple)

Returns whether an Entity has all given components.

For a more convenient tuple syntax, the macro @has_components is provided.

Example

has = has_components(world, entity, Val.((Position, Velocity)))

has_components(map::Map, entity::Entity)

Returns whether an Entity has all the Map's components.

Ark.@has_components — Macro

@has_components(world::World, entity::Entity, comp_types::Tuple)

Returns whether an Entity has all given components.

Macro version of has_components for more ergonomic component type tuples.

Example

has = @has_components(world, entity, (Position, Velocity))

Ark.set_components! — Function

set_components!(world::World, entity::Entity, values::Tuple)

Sets the given component values for an Entity. Types are inferred from the values. The entity must already have all these components.

Ark.add_components! — Function

add_components!(world::World, entity::Entity, values::Tuple)

Adds the given component values to an Entity. Types are inferred from the values.

add_components!(map::Map, entity::Entity, value::Tuple)

Adds the values of the Map's components to an Entity.

Ark.remove_components! — Function

remove_components!(world::World, entity::Entity, comp_types::Tuple)

Removes the given components from an Entity.

For a more convenient tuple syntax, the macro @remove_components! is provided.

Example

remove_components!(world, entity, Val.((Position, Velocity)))

remove_components!(map::Map, entity::Entity)

Removes the Map's components from an Entity.

Ark.@remove_components! — Macro

@remove_components!(world::World, entity::Entity, comp_types::Tuple)

Removes the given components from an Entity.

Macro version of remove_components! for ergonomic construction of component mappers.

Example

@remove_components!(world, entity, (Position, Velocity))

Ark.exchange_components! — Function

exchange_components!(world::World{CS,CT,N}, entity::Entity; add::Tuple, remove::Tuple)

Adds and removes components on an Entity. Types are inferred from the add values.

For a more convenient tuple syntax, the macro @exchange_components! is provided.

Example

exchange_components!(world, entity;
    add=(Health(100),),
    remove=Val.((Position, Velocity)),
)

Ark.@exchange_components! — Macro

@exchange_components!(world::World, entity::Entity, add::Tuple, remove::Tuple)

Removes the given components from an Entity.

Macro version of exchange_components! for more ergonomic component type tuples.

Example

@exchange_components!(world, entity,
    add=(Health(100),),
    remove=Val.((Position, Velocity)),
)

Queries

Queries are used to filter and process Entities with a certain set of Components.

Ark.Query — Type

Query

A query for components.

Ark.@Query — Macro

@Query(
    world::World,
    comp_types::Tuple,
    with::Tuple=(),
    without::Tuple=(),
    optional::Tuple=()
)

Creates a query.

Macro version of Query that allows ergonomic construction of queries using simulated keyword arguments.

Queries can be stored and re-used. However, query creation is fast (<20ns), so this is mandatory.

Arguments

world: The World instance to query.

comp_types::Tuple: Components the query filters for and provides access to. Must be a literal tuple like (Position, Velocity).

with::Tuple: Additional components the entities must have. Passed as with=(Health,).

without::Tuple: Components the entities must not have. Passed as without=(Altitude,).

optional::Tuple: Components that are optional in the query. Passed as optional=(Velocity,).

Example

for (entities, pos_column, vel_column) in @Query(world, (Position, Velocity))
    for i in eachindex(entities)
        pos = pos_column[i]
        vel = vel_column[i]
        pos_column[i] = Position(pos.x + vel.dx, pos.y + vel.dy)
    end
end

Ark.close! — Method

close!(q::Query)

Closes the query and unlocks the world.

Must be called if a query is not fully iterated.

Ark.Entities — Type

Entities

Archetype column for entities. Can be iterated and indexed like a Vector.

Used in query iteration.

Resources

Resources are singleton-like data structures that appear only once in a World and are not associated to an Entities.

Ark.get_resource — Function

get_resource(world::World, res_type::Type{T})

Get the resource of type T from the world.

Ark.has_resource — Function

has_resource(world::World, res_type::Type{T})

Check if a resource of type T is in the world.

Ark.add_resource! — Function

add_resource!(world::World, res::T)

Add the given resource to the world.

Ark.remove_resource! — Function

remove_resource!(world::World, res_type::Type{T})

Remove the resource of type T from the world.

Batch

Ark.Batch — Type

Batch

A batch iterator. This is returned from batch operations and serves for initializing newly added components.

Ark.close! — Method

close!(b::Batch)

Closes the batch iterator and unlocks the world.

Must be called if a batch is not fully iterated.

Index

Ark.zero_entity

Ark.Batch

Ark.Entities

Ark.Entity

Ark.Map

Ark.Query

Ark.World

Ark.add_components!

Ark.add_resource!

Ark.close!

Ark.close!

Ark.exchange_components!

Ark.get_components

Ark.get_resource

Ark.has_components

Ark.has_resource

Ark.is_alive

Ark.is_locked

Ark.is_zero

Ark.new_entities!

Ark.new_entity!

Ark.remove_components!

Ark.remove_entity!

Ark.remove_resource!

Ark.set_components!

Base.getindex

Base.setindex!

Ark.@Map

Ark.@Query

Ark.@exchange_components!

Ark.@get_components

Ark.@has_components

Ark.@new_entities!

Ark.@remove_components!